requirements.txt
Twisted==15.5.0
docopt==0.6.2
pyasn1==0.1.9
pycrypto==2.6.1
wsgiref==0.1.2
zope.interface==4.1.3
